:recycle: 페어룸 온보딩 모달 페이지 형태로 변경
