+++ frontend/src/App.tsx
@@ -1,4 +1,3 @@
-// import { useEffect } from 'react';
 import { useEffect } from 'react';
 import { createBrowserRouter, RouterProvider } from 'react-router-dom';
 
@@ -26,10 +25,12 @@ const queryClient = new QueryClient();
 
 const App = () => {
   const { setUserStatus } = useUserStatusStore();
+
   const checkUserStatus = async () => {
     const response = await getIsUserLoggedIn();
     setUserStatus(response.signedIn ? 'SIGNED_IN' : 'SIGNED_OUT');
   };
+
   useEffect(() => {
     if (process.env.NODE_ENV === 'production') {
       checkUserStatus();
@@ -53,6 +54,10 @@ const App = () => {
         {
           path: 'how-to-pair',
           element: <HowToPair />,
+        },
+        {
+          path: 'onboarding',
+          element: <PairRoomOnboarding />,
         },
         {
           path: 'room/:accessCode',
