+++ frontend/src/components/PairRoomOnboarding/PairRoleInput/PairRoleInput.styles.ts
@@ -2,18 +2,26 @@ import styled from 'styled-components';
 
 export const Layout = styled.div`
   display: flex;
-  flex: 1;
   flex-direction: column;
-  justify-content: space-between;
-
-  width: 100%;
-  padding: 4rem 0;
+  gap: 3rem;
 `;
 
-export const Container = styled.div`
+export const TitleContainer = styled.div`
   display: flex;
   flex-direction: column;
-  gap: 5rem;
+  gap: 1rem;
+`;
+
+export const Title = styled.div`
+  color: ${({ theme }) => theme.color.primary[700]};
+  font-size: ${({ theme }) => theme.fontSize.h5};
+  font-weight: ${({ theme }) => theme.fontWeight.medium};
+`;
+
+export const SubTitle = styled.div`
+  color: ${({ theme }) => theme.color.primary[600]};
+  font-size: ${({ theme }) => theme.fontSize.lg};
+  font-weight: ${({ theme }) => theme.fontWeight.normal};
 `;
 
 export const HeaderContainer = styled.div`
