+++ frontend/src/components/PairRoomOnboarding/StartMission/StartMission.styles.ts
