+++ frontend/src/components/PairRoomOnboarding/StartMission/StartMission.tsx
