+++ frontend/src/components/PairRoomOnboarding/TimerInput/TimerInput.styles.ts
@@ -2,19 +2,26 @@ import styled, { css } from 'styled-components';
 
 export const Layout = styled.div`
   display: flex;
-  flex: 1;
   flex-direction: column;
-  justify-content: space-between;
-  gap: 5rem;
-
-  width: 100%;
-  padding: 4rem 0;
+  gap: 3rem;
 `;
 
-export const Container = styled.div`
+export const TitleContainer = styled.div`
   display: flex;
   flex-direction: column;
-  gap: 5rem;
+  gap: 1rem;
+`;
+
+export const Title = styled.div`
+  color: ${({ theme }) => theme.color.primary[700]};
+  font-size: ${({ theme }) => theme.fontSize.h5};
+  font-weight: ${({ theme }) => theme.fontWeight.medium};
+`;
+
+export const SubTitle = styled.div`
+  color: ${({ theme }) => theme.color.primary[600]};
+  font-size: ${({ theme }) => theme.fontSize.lg};
+  font-weight: ${({ theme }) => theme.fontWeight.normal};
 `;
 
 export const HeaderContainer = styled.div`
