+++ frontend/src/pages/PairRoomOnboarding/PairRoomOnboarding.styles.ts
@@ -3,22 +3,34 @@ import styled from 'styled-components';
 export const Layout = styled.div`
   display: flex;
   justify-content: center;
-  align-items: flex-end;
 
   width: 100%;
-  height: calc(100vh - 7rem);
+  min-height: calc(100vh - 7rem);
 
   background-color: ${({ theme }) => theme.color.primary[100]};
+`;
+
+export const Title = styled.h2`
+  color: ${({ theme }) => theme.color.primary[800]};
+  font-size: ${({ theme }) => theme.fontSize.h3};
+  font-weight: ${({ theme }) => theme.fontWeight.semibold};
 `;
 
 export const Container = styled.div`
   display: flex;
   flex-direction: column;
+  gap: 5rem;
 
   width: 60%;
-  height: 95%;
   padding: 4rem;
-  border-radius: 4rem 4rem 0 0;
 
   background-color: ${({ theme }) => theme.color.black[10]};
+`;
+
+export const ButtonWrapper = styled.div`
+  display: flex;
+  justify-content: center;
+
+  width: 100%;
+  margin-top: 5rem;
 `;
